+++ tests/TestCase/Engine/Cache/MemcachedEngineTest.php
@@ -53,7 +53,7 @@ class MemcachedEngineTest extends \PHPUnit\Framework\TestCase
     public function testSet()
     {
         $cache = $this->engine();
-        $this->assertTrue($cache->set('foo', 'bar'));
+        $this->assertTrue($cache->write('foo', 'bar'));
         $this->assertEquals('bar', $cache->memcached()->get('origin_foo'));
     }
     /**
@@ -62,9 +62,9 @@ class MemcachedEngineTest extends \PHPUnit\Framework\TestCase
     public function testGet()
     {
         $cache = $this->engine();
-        $this->assertFalse($cache->get('foo'));
-        $cache->set('foo', 'bar');
-        $this->assertEquals('bar', $cache->get('foo'));
+        $this->assertFalse($cache->read('foo'));
+        $cache->write('foo', 'bar');
+        $this->assertEquals('bar', $cache->read('foo'));
     }
     /**
      * @depends testSet
@@ -72,9 +72,9 @@ class MemcachedEngineTest extends \PHPUnit\Framework\TestCase
     public function testHas()
     {
         $cache = $this->engine();
-        $this->assertFalse($cache->has('foo'));
-        $cache->set('foo', 'bar');
-        $this->assertTrue($cache->has('foo'));
+        $this->assertFalse($cache->exists('foo'));
+        $cache->write('foo', 'bar');
+        $this->assertTrue($cache->exists('foo'));
     }
     /**
      * @depends testHas
@@ -82,11 +82,11 @@ class MemcachedEngineTest extends \PHPUnit\Framework\TestCase
     public function testDelete()
     {
         $cache = $this->engine();
-        $cache->set('foo', 'bar');
-        $this->assertTrue($cache->has('foo'));
+        $cache->write('foo', 'bar');
+        $this->assertTrue($cache->exists('foo'));
         $this->assertTrue($cache->delete('foo'));
         
-        $this->assertFalse($cache->has('foo'));
+        $this->assertFalse($cache->exists('foo'));
         $this->assertFalse($cache->delete('foo'));
     }
     /**
@@ -95,11 +95,11 @@ class MemcachedEngineTest extends \PHPUnit\Framework\TestCase
     public function testClear()
     {
         $cache = $this->engine();
-        $cache->set('foo', 'bar');
-        $cache->set('bar', 'foo');
+        $cache->write('foo', 'bar');
+        $cache->write('bar', 'foo');
         $this->assertTrue($cache->clear());
-        $this->assertFalse($cache->has('foo'));
-        $this->assertFalse($cache->has('bar'));
+        $this->assertFalse($cache->exists('foo'));
+        $this->assertFalse($cache->exists('bar'));
     }
     /**
      * @depends testSet
@@ -107,7 +107,7 @@ class MemcachedEngineTest extends \PHPUnit\Framework\TestCase
     public function testIncrement()
     {
         $cache = $this->engine();
-        $cache->set('counter', 100);
+        $cache->write('counter', 100);
         $this->assertEquals(101, $cache->increment('counter'));
         $this->assertEquals(110, $cache->increment('counter', 9));
     }
@@ -117,7 +117,7 @@ class MemcachedEngineTest extends \PHPUnit\Framework\TestCase
     public function testDecrement()
     {
         $cache = $this->engine();
-        $cache->set('counter', 110);
+        $cache->write('counter', 110);
         $this->assertEquals(109, $cache->decrement('counter'));
         $this->assertEquals(100, $cache->decrement('counter', 9));
     }