+++ tests/TestCase/Engine/Cache/RedisEngineTest.php
@@ -54,7 +54,7 @@ class RedisEngineTest extends \PHPUnit\Framework\TestCase
     public function testSet()
     {
         $cache = $this->engine();
-        $this->assertTrue($cache->set('foo', 'bar'));
+        $this->assertTrue($cache->write('foo', 'bar'));
         $this->assertEquals('bar', $cache->redis()->get('origin_foo'));
     }
     /**
@@ -63,9 +63,9 @@ class RedisEngineTest extends \PHPUnit\Framework\TestCase
     public function testGet()
     {
         $cache = $this->engine();
-        $this->assertFalse($cache->get('foo'));
-        $cache->set('foo', 'bar');
-        $this->assertEquals('bar', $cache->get('foo'));
+        $this->assertFalse($cache->read('foo'));
+        $cache->write('foo', 'bar');
+        $this->assertEquals('bar', $cache->read('foo'));
     }
     /**
      * @depends testSet
@@ -73,9 +73,9 @@ class RedisEngineTest extends \PHPUnit\Framework\TestCase
     public function testHas()
     {
         $cache = $this->engine();
-        $this->assertFalse($cache->has('foo'));
-        $cache->set('foo', 'bar');
-        $this->assertTrue($cache->has('foo'));
+        $this->assertFalse($cache->exists('foo'));
+        $cache->write('foo', 'bar');
+        $this->assertTrue($cache->exists('foo'));
     }
     /**
      * @depends testHas
@@ -83,11 +83,11 @@ class RedisEngineTest extends \PHPUnit\Framework\TestCase
     public function testDelete()
     {
         $cache = $this->engine();
-        $cache->set('foo', 'bar');
-        $this->assertTrue($cache->has('foo'));
+        $cache->write('foo', 'bar');
+        $this->assertTrue($cache->exists('foo'));
         $this->assertTrue($cache->delete('foo'));
         
-        $this->assertFalse($cache->has('foo'));
+        $this->assertFalse($cache->exists('foo'));
         $this->assertFalse($cache->delete('foo'));
     }
     /**
@@ -96,11 +96,11 @@ class RedisEngineTest extends \PHPUnit\Framework\TestCase
     public function testClear()
     {
         $cache = $this->engine();
-        $cache->set('foo', 'bar');
-        $cache->set('bar', 'foo');
+        $cache->write('foo', 'bar');
+        $cache->write('bar', 'foo');
         $this->assertTrue($cache->clear());
-        $this->assertFalse($cache->has('foo'));
-        $this->assertFalse($cache->has('bar'));
+        $this->assertFalse($cache->exists('foo'));
+        $this->assertFalse($cache->exists('bar'));
     }
     /**
      * @depends testSet
@@ -108,7 +108,7 @@ class RedisEngineTest extends \PHPUnit\Framework\TestCase
     public function testIncrement()
     {
         $cache = $this->engine();
-        $cache->set('counter', 100);
+        $cache->write('counter', 100);
         $this->assertEquals(101, $cache->increment('counter'));
         $this->assertEquals(110, $cache->increment('counter', 9));
     }
@@ -118,7 +118,7 @@ class RedisEngineTest extends \PHPUnit\Framework\TestCase
     public function testDecrement()
     {
         $cache = $this->engine();
-        $cache->set('counter', 110);
+        $cache->write('counter', 110);
         $this->assertEquals(109, $cache->decrement('counter'));
         $this->assertEquals(100, $cache->decrement('counter', 9));
     }
@@ -144,7 +144,7 @@ class RedisEngineTest extends \PHPUnit\Framework\TestCase
             'persistent' => 'persisten-id'
         ]);
     
-        $redis ->set('counter', 100);
+        $redis ->write('counter', 100);
         $this->assertEquals(101, $redis->increment('counter'));
     }
 }